+++ packages/ext/entrypoints/background/llm.ts
@@ -6,30 +6,34 @@ import { streamText } from "ai";
 
 export function registerLLMMessages() {
   onConnectMessage('streamTextViaConnect', async ({ messages, modelConfig }, { chunk, chunkEnd, complete, error, markReturn, resolve }) => {
-    const options:{temperature?:number, topP?:number,topK?:number, frequencyPenalty?:number, presencePenalty?:number, maxTokens?:number}={}
-    if(modelConfig.temperature){
-      options.temperature=modelConfig.temperature
+    const options: { temperature?: number, topP?: number, topK?: number, frequencyPenalty?: number, presencePenalty?: number, maxTokens?: number } = {}
+    if (modelConfig.temperature) {
+      options.temperature = modelConfig.temperature
     }
-    if(modelConfig.topP){
-      options.topP=modelConfig.topP
+    if (modelConfig.topP) {
+      options.topP = modelConfig.topP
     }
-    if(modelConfig.topK){
-      options.topK=modelConfig.topK
+    if (modelConfig.topK) {
+      options.topK = modelConfig.topK
     }
-    if(modelConfig.frequencyPenalty){
-      options.frequencyPenalty=modelConfig.frequencyPenalty
+    if (modelConfig.frequencyPenalty) {
+      options.frequencyPenalty = modelConfig.frequencyPenalty
     }
-    if(modelConfig.presencePenalty){
-      options.presencePenalty=modelConfig.presencePenalty
+    if (modelConfig.presencePenalty) {
+      options.presencePenalty = modelConfig.presencePenalty
     }
-    if(modelConfig.maxTokens){
-      options.maxTokens=modelConfig.maxTokens
+    if (modelConfig.maxTokens) {
+      options.maxTokens = modelConfig.maxTokens
     }
     try {
       const { usage, textStream } = streamText({
+        ...options,
         messages: messages,
         model: createVercelModel(modelConfig),
-        ...options
+        onError(e: any) {
+          console.error('[streamTextViaConnect]streamText.onError:', e)
+          error(unpackStreamTextError(e))
+        }
       })
       markReturn([
         { key: 'textStream', type: 'chunk' },
@@ -51,15 +55,26 @@ export function registerLLMMessages() {
       chunkEnd('textStream')
 
     } catch (e: any) {
-      console.log('[streamTextViaConnect]occurs error', e,e instanceof TypeError)
-      if(e instanceof TypeError){
-        error({name: e.name, message: e.message})
-        return 
-      }
-      
-      
-      error(e)
+      console.error('[streamTextViaConnect]catch error:', e)
+      error(unpackStreamTextError(e))
     }
+    console.debug('[streamTextViaConnect]complete')
 
   })
+}
+
+
+function unpackStreamTextError(e: any) {
+  /*
+    * for  @ai-sdk/anthropic, it's a {error:Error} object
+  */
+  if (Object.keys(e).length === 1 && e.hasOwnProperty('error')) {
+    e = e.error
+  }
+  if (e instanceof TypeError) {
+    e = { name: e.name, message: e.message }
+    return
+  }
+
+  return e;
 }